requirements.txt
flask
flask-cors
flask-migrate
flask-sqlalchemy
python-dotenv
flask-swagger
flask-admin


